refactor: move generic utils to src
